Fix cost differ if varying spread is used or not >Try to debug bad TLM gradients when varying spread is used

+++ build_gamma.sh
@@ -1,16 +1,48 @@
+#!/bin/bash
 #Parse input arguments
+
 DEBUG=0
+F90BUILD=0
+PYWRAPPERS=0
+SHAREDLIB=0
+INSTALL=0
+CLEAN=0
+
 i=1
 until [ $# = 0 ]
 do
 case $1 in
+    --all)
+    shift
+    F90BUILD=1
+    PYWRAPPERS=1
+    SHAREDLIB=1
+    INSTALL=1
+    CLEAN=1
+    ;;
     --debug)
     shift
     DEBUG=1
     ;;
-    --build_only)
+    --f90build)
     shift
-    BUILD_ONLY=1
+    F90BUILD=1
+    ;;
+    --pywrappers)
+    shift
+    PYWRAPPERS=1
+    ;;
+    --sharedlib)
+    shift
+    SHAREDLIB=1
+    ;;
+    --install)
+    shift
+    INSTALL=1
+    ;;
+    --clean)
+    shift
+    CLEAN=1
     ;;
     *)
     echo "Unknow option $1 ..."
@@ -36,11 +68,15 @@ src_f90wrap=$src_dir/f90wrap/
 wrapped_lib_name="libfgamma"
 build_dir=$(pwd)/build/
 
-rm -r $src_py/gamma_routing_model/$wrapped_lib_name
-rm -r $build_dir
+if [ $CLEAN -eq 1 ] ; then
+    rm -r $src_py/gamma_routing_model/$wrapped_lib_name
+    rm -r $build_dir
+fi
 
 mkdir -p $build_dir
 cd $build_dir
+#creation d'un module python
+mkdir -p $wrapped_lib_name
 
 cp $src_f90/*.f90 .
 cp $src_f77*.f .
@@ -66,42 +102,53 @@ else
     f_flags="-fPIC -O3 -march=native -funroll-loops -fallow-argument-mismatch"
 fi
 
+if [ $PYWRAPPERS -eq 1 ] ; then
+    #Make the python wrappers and create a python module libfgamma
+    f90wrap -k "kind_map" -m "$wrapped_lib_name" $f90files_wrap
+    
+    #change how the lib is imported
+    sed -i "s/^\(import \_${wrapped_lib_name}\)/from \. \1/g" ${wrapped_lib_name}.py
+    mv ./${wrapped_lib_name}.py ${wrapped_lib_name}/.
+fi
 
-#Make the python wrappers and create a python module libfgamma
-f90wrap -k "kind_map" -m "$wrapped_lib_name" $f90files_wrap
+if [ $F90BUILD -eq 1 ] ; then
+    #compilation des sources et créations des objets .o
+    gcc $c_flags -c -g -fPIC $cfiles
+    gfortran $f_flags -c $f77files
+    gfortran $f_flags -c $f90files
 
-#order matters 
-f90wrapped_files=""
-for file in $f90files_wrap
-do
-    f90wrapped_files+=" f90wrap_${file}"
-done
+    #compilation of the main prog
+    gfortran $f90mainprog *.o -o $(basename f90mainprog)
+fi
 
-#compilation des sources et créations des objets .o
-gcc $c_flags -c -g -fPIC $cfiles
-gfortran $f_flags -c $f77files
-gfortran $f_flags -c $f90files
+if [ $SHAREDLIB -eq 1 ] ; then
 
-#compilation of the main prog
-gfortran $f90mainprog *.o -o $(basename f90mainprog)
+    #order matters 
+    f90wrapped_files=""
+    for file in $f90files_wrap
+    do
+        f90wrapped_files+=" f90wrap_${file}"
+    done
+    
+    #Création de la librairie partagée
+    FFLAGS="-I$(pwd)" FC=gfortran f2py -c --f90flags="$f_flags" -m "_${wrapped_lib_name}" $f90wrapped_files *.o
+    
+    #create the __init__.py module file
+    echo "from .libfgamma import Mod_Gamma_Interface, Mod_Gamma_Routing_Setup, Mod_Gamma_Routing_Mesh, Mod_Gamma_Routing_Parameters, Mod_Gamma_Routing_Results, Mod_Gamma_Routing_States" > ${wrapped_lib_name}/__init__.py
 
-#Création de la librairie partagée
-FFLAGS="-I$(pwd)" FC=gfortran f2py -c --f90flags="$f_flags" -m "_${wrapped_lib_name}" $f90wrapped_files *.o
+    #move all component inside the module directories and move it inside the source module
+    mv _${wrapped_lib_name}* ${wrapped_lib_name}/.
+fi
 
-#creation d'un module python
-mkdir -p $wrapped_lib_name
+#Before install the gamma module, copy the wrappers and the shared lib module in the source code
+cp -u -r ./${wrapped_lib_name} $src_py/gamma_routing_model/.
 
-#change how the lib is imported
-sed -i "s/^\(import \_${wrapped_lib_name}\)/from \. \1/g" ${wrapped_lib_name}.py
-#create the __init__.py module file
-echo "from .libfgamma import Mod_Gamma_Interface, Mod_Gamma_Routing_Setup, Mod_Gamma_Routing_Mesh, Mod_Gamma_Routing_Parameters, Mod_Gamma_Routing_Results, Mod_Gamma_Routing_States" > ${wrapped_lib_name}/__init__.py
-
-#move all component inside the module directories and move it inside the source module
-mv ./${wrapped_lib_name}.py ${wrapped_lib_name}/.
-mv _${wrapped_lib_name}* ${wrapped_lib_name}/.
-mv ./${wrapped_lib_name} $src_py/gamma_routing_model
-
-#intall the module with editable mode
-cd $src_dir
-pip install -e ./py
-
+if [ $INSTALL -eq 1 ] ; then
+    cd $src_dir
+    if [ DEBUG -eq 1 ] ; then
+        #intall the module with editable mode
+        pip install -e ./py
+    else
+        pip install ./py
+    fi
+fi
